refactor: 루트 레이아웃에서 프로바이더 통합 관리
