--- src/app/(causw)/board/Provider.tsx
@@ -1,27 +1,0 @@
-'use client';
-
-import { ReactNode, useState } from 'react';
-
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-
-import { ToastWithMax } from '@/fsd_shared';
-
-export const Provider = ({ children }: { children: ReactNode }) => {
-  const [queryClient] = useState(
-    () =>
-      new QueryClient({
-        defaultOptions: {
-          queries: {
-            staleTime: 60 * 1000,
-            refetchOnWindowFocus: false,
-          },
-        },
-      }),
-  );
-  return (
-    <QueryClientProvider client={queryClient}>
-      <ToastWithMax />
-      {children}
-    </QueryClientProvider>
-  );
-};
