--- src/app/(causw)/board/layout.tsx
@@ -1,9 +1,0 @@
-import { Provider } from './Provider';
-
-const Layout = ({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) => <Provider>{children}</Provider>;
-
-export default Layout;
