--- src/app/(causw)/setting/management/attendance/detail/layout.tsx
@@ -1,13 +1,0 @@
-'use client';
-
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-
-const queryClient = new QueryClient();
-
-const QueryClientLayout = ({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) => <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>;
-
-export default QueryClientLayout;
