--- src/app/(causw)/setting/management/user/[state]/[admissionId]/layout.tsx
@@ -1,16 +1,0 @@
-'use client';
-
-import { ToastWithMax } from '@/shared';
-
-const toastLayout = ({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) => (
-  <>
-    <ToastWithMax />
-    {children}
-  </>
-);
-
-export default toastLayout;
