--- src/app/(causw)/setting/personal-info/update-academic-record/layout.tsx
@@ -1,16 +1,0 @@
-'use client';
-
-import { ToastWithMax } from '@/shared';
-
-const Layout = ({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) => (
-  <>
-    <ToastWithMax />
-    {children}
-  </>
-);
-
-export default Layout;
